refactor: :recycle: style 파일 하나로 합침 (굳이 분리할 이유를 모르겠어서)
