--- frontend/src/pages/ChatPage/ChatModal/ChatInfoModalBody.styles.ts
@@ -1,9 +1,0 @@
-import { css } from '@emotion/css';
-
-export const chatInfoModalBodyStyle = css`
-  width: 100%;
-  height: 65%;
-  overflow-x: hidden;
-  overflow-y: scroll;
-  border-bottom: 1px solid black; // TODO: 상수화
-`;
